Add wasm-bindgen to the rustc build
This commit is targeted at enabling by default the `#[wasm_bindgen]` attribute
for Rust code. This involved a few primary pieces to be available on the
builder:

* First the runtime support of the `wasm_bindgen` crate and the
  `#[wasm_bindgen]` macro needed to be available when compiling code.
* Next the `wasm-bindgen` CLI tool needed to be available and automatically run
  whenever a wasm file is generated with an appropriate section name.
* Finally the generated JS as well as the generated wasm needed to be sent back
  to the client

The `/rustc` has been updated with all these steps and transitioned to `async`
and `await` at the same time to make it a bit clearer to see what's going on.
The slug generation has also been updated to primarily go through Docker to
ensure a nicely containerized build. I've deployed this current slug to
https://young-sierra-27135.herokuapp.com/ for inspection as well.